pyproject.toml
[tool.poetry]
name = "hscode-info"
version = "0.1.0"
description = ""
authors = ["jaeupsu <hyeongs2323@gmail.com>"]
readme = "README.md"
packages = [{include = "hscode_info"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
